Make chat_id the source of truth for chat events; split ChatManager module
- Split events.py into schema-only (events.py) + ChatManager (chat_manager.py).
- Chat lifecycle events now carry the stable chat_id (chat.get_id()) instead
  of the RTC room_id, which is dropped from the schema/ChatEvent entirely.
  room_id stays internal to ChatManager RTC resolution and to YChat (path
  recovery). chat_id is captured before freeing a model so close/delete events
  still carry it, and the RTC opened event reads it after get_document loads
  content (never mints a premature id).
- Frontend: model.ready now resolves with the chat id (Promise<string>), and
  the WS path no longer mints a local UUID fallback -- the id is exactly the
  server's chat.get_id() (from the connection frame), guaranteeing the
  frontend id matches the backend.
